requirements.txt
asgiref==3.8.1
Django==5.0.6
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
drf-spectacular
